Reorganize so that database types and data are grouped together in their own files.
