--- src/data/DatabaseCategory.ts
@@ -1,21 +1,0 @@
-enum databaseCategory {
-  SEQUENCE = 'SEQUENCE',
-  STRUCTURE = 'STRUCTURE',
-  INTERACTION = 'INTERACTION',
-  CHEMISTRY = 'CHEMISTRY',
-  FAMILY = 'FAMILY',
-  PTM = 'PTM',
-  POLYMORPHISM = 'POLYMORPHISM',
-  GEL = 'GEL',
-  PROTEOMIC = 'PROTEOMIC',
-  PROTOCOL = 'PROTOCOL',
-  GENOME = 'GENOME',
-  ORGANISM = 'ORGANISM',
-  PHYLOGENOMIC = 'PHYLOGENOMIC',
-  PATHWAY = 'PATHWAY',
-  EXPRESSION = 'EXPRESSION',
-  DOMAIN = 'DOMAIN',
-  OTHER = 'OTHER',
-}
-
-export default databaseCategory;
