+++ src/model/types/EntrySection.ts
