+++ src/model/types/databaseTypes.ts
@@ -1,0 +1,200 @@
+export enum DatabaseCategory {
+  SEQUENCE = 'SEQUENCE',
+  STRUCTURE = 'STRUCTURE',
+  INTERACTION = 'INTERACTION',
+  CHEMISTRY = 'CHEMISTRY',
+  FAMILY = 'FAMILY',
+  PTM = 'PTM',
+  POLYMORPHISM = 'POLYMORPHISM',
+  GEL = 'GEL',
+  PROTEOMIC = 'PROTEOMIC',
+  PROTOCOL = 'PROTOCOL',
+  GENOME = 'GENOME',
+  ORGANISM = 'ORGANISM',
+  PHYLOGENOMIC = 'PHYLOGENOMIC',
+  PATHWAY = 'PATHWAY',
+  EXPRESSION = 'EXPRESSION',
+  DOMAIN = 'DOMAIN',
+  OTHER = 'OTHER',
+}
+
+type AttributesItem = {
+  name: string;
+  xmlTag: string;
+  uriLink?: string;
+};
+
+type DatabaseInfo = {
+  name: string;
+  displayName?: string;
+  category: string;
+  uriLink: string;
+  attributes?: AttributesItem[];
+};
+
+export type DatabaseToDatabaseInfo = { [key in Database]: DatabaseInfo };
+
+export enum Database {
+  EMBL = 'EMBL',
+  CCDS = 'CCDS',
+  PIR = 'PIR',
+  RefSeq = 'RefSeq',
+  UniGene = 'UniGene',
+  PDB = 'PDB',
+  PDBsum = 'PDBsum',
+  DisProt = 'DisProt',
+  ProteinModelPortal = 'ProteinModelPortal',
+  SMR = 'SMR',
+  BioGrid = 'BioGrid',
+  ComplexPortal = 'ComplexPortal',
+  CORUM = 'CORUM',
+  DIP = 'DIP',
+  ELM = 'ELM',
+  IntAct = 'IntAct',
+  MINT = 'MINT',
+  STRING = 'STRING',
+  BindingDB = 'BindingDB',
+  ChEMBL = 'ChEMBL',
+  DrugBank = 'DrugBank',
+  GuidetoPHARMACOLOGY = 'GuidetoPHARMACOLOGY',
+  SwissLipids = 'SwissLipids',
+  Allergome = 'Allergome',
+  CAZy = 'CAZy',
+  ESTHER = 'ESTHER',
+  IMGT_GENE_DB = 'IMGT_GENE-DB',
+  MEROPS = 'MEROPS',
+  MoonDB = 'MoonDB',
+  MoonProt = 'MoonProt',
+  mycoCLAP = 'mycoCLAP',
+  PeroxiBase = 'PeroxiBase',
+  REBASE = 'REBASE',
+  TCDB = 'TCDB',
+  UniLectin = 'UniLectin',
+  CarbonylDB = 'CarbonylDB',
+  DEPOD = 'DEPOD',
+  GlyConnect = 'GlyConnect',
+  iPTMnet = 'iPTMnet',
+  PhosphoSitePlus = 'PhosphoSitePlus',
+  SwissPalm = 'SwissPalm',
+  UniCarbKB = 'UniCarbKB',
+  BioMuta = 'BioMuta',
+  DMDM = 'DMDM',
+  dbSNP = 'dbSNP',
+  COMPLUYEAST_2DPAGE = 'COMPLUYEAST-2DPAGE',
+  DOSAC_COBS_2DPAGE = 'DOSAC-COBS-2DPAGE',
+  OGP = 'OGP',
+  REPRODUCTION_2DPAGE = 'REPRODUCTION-2DPAGE',
+  SWISS_2DPAGE = 'SWISS-2DPAGE',
+  UCD_2DPAGE = 'UCD-2DPAGE',
+  World_2DPAGE = 'World-2DPAGE',
+  EPD = 'EPD',
+  MaxQB = 'MaxQB',
+  PaxDb = 'PaxDb',
+  PeptideAtlas = 'PeptideAtlas',
+  PRIDE = 'PRIDE',
+  ProMEX = 'ProMEX',
+  ProteomicsDB = 'ProteomicsDB',
+  TopDownProteomics = 'TopDownProteomics',
+  jPOST = 'jPOST',
+  DNASU = 'DNASU',
+  Ensembl = 'Ensembl',
+  EnsemblBacteria = 'EnsemblBacteria',
+  EnsemblFungi = 'EnsemblFungi',
+  EnsemblMetazoa = 'EnsemblMetazoa',
+  EnsemblPlants = 'EnsemblPlants',
+  EnsemblProtists = 'EnsemblProtists',
+  GeneDB = 'GeneDB',
+  GeneID = 'GeneID',
+  Gramene = 'Gramene',
+  KEGG = 'KEGG',
+  PATRIC = 'PATRIC',
+  UCSC = 'UCSC',
+  VectorBase = 'VectorBase',
+  WBParaSite = 'WBParaSite',
+  ArachnoServer = 'ArachnoServer',
+  Araport = 'Araport',
+  CGD = 'CGD',
+  ConoServer = 'ConoServer',
+  CTD = 'CTD',
+  dictyBase = 'dictyBase',
+  DisGeNET = 'DisGeNET',
+  EchoBASE = 'EchoBASE',
+  EcoGene = 'EcoGene',
+  euHCVdb = 'euHCVdb',
+  EuPathDB = 'EuPathDB',
+  FlyBase = 'FlyBase',
+  GeneCards = 'GeneCards',
+  GeneReviews = 'GeneReviews',
+  H_InvDB = 'H-InvDB',
+  HGNC = 'HGNC',
+  HPA = 'HPA',
+  LegioList = 'LegioList',
+  Leproma = 'Leproma',
+  MaizeGDB = 'MaizeGDB',
+  MalaCards = 'MalaCards',
+  MGI = 'MGI',
+  MIM = 'MIM',
+  neXtProt = 'neXtProt',
+  OpenTargets = 'OpenTargets',
+  Orphanet = 'Orphanet',
+  PharmGKB = 'PharmGKB',
+  PomBase = 'PomBase',
+  PseudoCAP = 'PseudoCAP',
+  RGD = 'RGD',
+  SGD = 'SGD',
+  TAIR = 'TAIR',
+  TubercuList = 'TubercuList',
+  VGNC = 'VGNC',
+  WormBase = 'WormBase',
+  Xenbase = 'Xenbase',
+  ZFIN = 'ZFIN',
+  eggNOG = 'eggNOG',
+  GeneTree = 'GeneTree',
+  HOGENOM = 'HOGENOM',
+  HOVERGEN = 'HOVERGEN',
+  InParanoid = 'InParanoid',
+  KO = 'KO',
+  OMA = 'OMA',
+  OrthoDB = 'OrthoDB',
+  PhylomeDB = 'PhylomeDB',
+  TreeFam = 'TreeFam',
+  BioCyc = 'BioCyc',
+  BRENDA = 'BRENDA',
+  Reactome = 'Reactome',
+  SABIO_RK = 'SABIO-RK',
+  SignaLink = 'SignaLink',
+  SIGNOR = 'SIGNOR',
+  UniPathway = 'UniPathway',
+  ChiTaRS = 'ChiTaRS',
+  EvolutionaryTrace = 'EvolutionaryTrace',
+  GeneWiki = 'GeneWiki',
+  GenomeRNAi = 'GenomeRNAi',
+  PMAP_CutDB = 'PMAP-CutDB',
+  PRO = 'PRO',
+  Bgee = 'Bgee',
+  CleanEx = 'CleanEx',
+  CollecTF = 'CollecTF',
+  ExpressionAtlas = 'ExpressionAtlas',
+  Genevisible = 'Genevisible',
+  CDD = 'CDD',
+  Gene3D = 'Gene3D',
+  HAMAP = 'HAMAP',
+  InterPro = 'InterPro',
+  PANTHER = 'PANTHER',
+  Pfam = 'Pfam',
+  PIRSF = 'PIRSF',
+  PRINTS = 'PRINTS',
+  ProDom = 'ProDom',
+  SFLD = 'SFLD',
+  SMART = 'SMART',
+  SUPFAM = 'SUPFAM',
+  TIGRFAMs = 'TIGRFAMs',
+  PROSITE = 'PROSITE',
+  GO = 'GO',
+  Proteomes = 'Proteomes',
+  GlycoProtDB = 'GlycoProtDB',
+  ProtClustDB = 'ProtClustDB',
+  PathwayCommons = 'PathwayCommons',
+}
+
+export default Database;
